Leasing: 4kv and 10kv converted to float instead of integer while sending to udb

+++ kcell-process-app/src/main/java/kz/kcell/flow/leasing/CreateCandidate.java
@@ -1105,14 +1105,14 @@ public class CreateCandidate implements JavaDelegate {
                         ARTEFACT_RR_POWER_PreparedStatement.setNull(i++, Types.BIGINT);
                     }
                     if (res_4kv != null) {
-                        ARTEFACT_RR_POWER_PreparedStatement.setLong(i++, Integer.parseInt(res_4kv)); //RES_4KV
+                        ARTEFACT_RR_POWER_PreparedStatement.setFloat(i++, Float.parseFloat(res_4kv)); //RES_4KV
                     } else {
-                        ARTEFACT_RR_POWER_PreparedStatement.setNull(i++, Types.BIGINT);
+                        ARTEFACT_RR_POWER_PreparedStatement.setNull(i++, Types.FLOAT);
                     }
                     if (res_10kv != null) {
-                        ARTEFACT_RR_POWER_PreparedStatement.setLong(i++, Integer.parseInt(res_10kv)); //RES_10KV
+                        ARTEFACT_RR_POWER_PreparedStatement.setFloat(i++, Integer.parseFloat(res_10kv)); //RES_10KV
                     } else {
-                        ARTEFACT_RR_POWER_PreparedStatement.setNull(i++, Types.BIGINT);
+                        ARTEFACT_RR_POWER_PreparedStatement.setNull(i++, Types.FLOAT);
                     }
